+++ tests/unit/test_init.py
@@ -1,7 +1,5 @@
 """Tests for generic parser."""
 import os
-from pathlib import Path
-import email
 
 import pytest
 
@@ -9,11 +7,7 @@ from circuit_maintenance_parser import (
     init_provider,
     get_provider_class,
     get_provider_class_from_sender,
-    init_data_raw,
-    init_data_email,
-    init_data_emailmessage,
 )
-from circuit_maintenance_parser.data import NotificationData
 from circuit_maintenance_parser.errors import NonexistentProviderError
 from circuit_maintenance_parser.provider import (
     GenericProvider,
@@ -25,35 +19,6 @@ from circuit_maintenance_parser.provider import (
 
 
 dir_path = os.path.dirname(os.path.realpath(__file__))
-
-
-def test_init_data_raw():
-    """Test the init_data_raw function."""
-    data = init_data_raw("my_type", b"my_content")
-    assert isinstance(data, NotificationData)
-    assert len(data.data_parts) == 1
-    assert data.data_parts[0].type == "my_type"
-    assert data.data_parts[0].content == b"my_content"
-
-
-def test_init_data_email():
-    """Test the email data load."""
-    with open(Path(dir_path, "data", "email", "test_sample_message.eml"), "rb") as email_file:
-        email_raw_data = email_file.read()
-    data = init_data_email(email_raw_data)
-    assert isinstance(data, NotificationData)
-    assert len(data.data_parts) == 5
-
-
-def test_init_data_emailmessage():
-    """Test the emailmessage data load."""
-    with open(Path(dir_path, "data", "email", "test_sample_message.eml"), "rb") as email_file:
-        email_raw_data = email_file.read()
-    raw_email_string = email_raw_data.decode("utf-8")
-    email_message = email.message_from_string(raw_email_string)
-    data = init_data_emailmessage(email_message)
-    assert isinstance(data, NotificationData)
-    assert len(data.data_parts) == 5
 
 
 @pytest.mark.parametrize(